jadn/requirements.txt
cbor2
dicttoxml
jsonschema
pyyaml
xmltodict